HealthCare Differential Diagnosis (DDx)/requirements.txt
openai==0.27.7
streamlit==1.22.0
streamlit_audiorecorder==0.0.2
